Removed Figure Visualization
At several points in the script, it looks like the original author included code to visualize various correlation matrices. Since the goal of this code is to loop over subject, and we will not be sitting there looking at each visualization, I removed all code having to do with creating MATLAB figures and visualizing matrices.