Add files via upload
We get this error:

Analysis: FAME_categorymodel_ret_hrf
Gathering Data...
Subject: 20y396
Run: 1
ERROR ON: 20y396>> 

It's referring to the catch error under the "Run the SPM job" .... I just don't know what that error is referring to. 

I had to comment out your recently added error catch (around LINE 89) ... it kept saying that the error being defined there was screwing up the later error definition at LINE 119. It would be good tho if we can get both of these errors to work correctly if you think that would be helpful.

Right now on line 21, it calls an nii. But when we look at our files on ANVIL, they are hdr and img files. We tried running this script with img and still had the same problem. Which should it be? 

The transfer of our files from ANVIL to ACI has been delayed two weeks, just FYI.